deploy/hf-space/requirements.txt
gradio==5.22.0
huggingface_hub==0.30.2
tensorflow-cpu==2.16.1
pillow==10.4.0
numpy==1.26.4
